+++ SPLASH/src/game/states/GameState.cpp
@@ -2,28 +2,7 @@
 #include "imgui.h"
 #include "Sail/entities/ECS.h"
 #include "Sail/entities/components/Components.h"
-#include "Sail/entities/systems/candles/CandleSystem.h"
-#include "Sail/entities/systems/entityManagement/EntityAdderSystem.h"
-#include "Sail/entities/systems/entityManagement/EntityRemovalSystem.h"
-#include "Sail/entities/systems/lifetime/LifeTimeSystem.h"
-#include "Sail/entities/systems/light/LightSystem.h"
-#include "Sail/entities/systems/Gameplay/ai/AiSystem.h"
-#include "Sail/entities/systems/Gameplay/GunSystem.h"
-#include "Sail/entities/systems/Gameplay/ProjectileSystem.h"
-#include "Sail/entities/systems/Graphics/AnimationSystem.h"
-#include "Sail/entities/systems/LevelGeneratorSystem/LevelGeneratorSystem.h"
-#include "Sail/entities/systems/physics/OctreeAddRemoverSystem.h"
-#include "Sail/entities/systems/physics/UpdateBoundingBoxSystem.h"
-#include "Sail/entities/systems/prepareUpdate/PrepareUpdateSystem.h"
-#include "Sail/entities/systems/input/GameInputSystem.h"
-#include "Sail/entities/systems/network/NetworkReceiverSystem.h"
-#include "Sail/entities/systems/network/NetworkSenderSystem.h"
-#include "Sail/entities/systems/Audio/AudioSystem.h"
-#include "Sail/entities/systems/render/RenderSystem.h"
-#include "Sail/entities/systems/physics/MovementSystem.h"
-#include "Sail/entities/systems/physics/MovementPostCollisionSystem.h"
-#include "Sail/entities/systems/physics/CollisionSystem.h"
-#include "Sail/entities/systems/physics/SpeedLimitSystem.h"
+#include "Sail/entities/systems/Systems.h"
 #include "Sail/ai/states/AttackingState.h"
 #include "Sail/ai/states/FleeingState.h"
 #include "Sail/ai/states/SearchingState.h"
