+++ SPLASH/src/game/states/GameState.h
@@ -1,28 +1,6 @@
 #pragma once
-
 #include "Sail.h"
-
-class AiSystem;
-class AnimationSystem;
-class CandleSystem;
-class EntityAdderSystem;
-class EntityRemovalSystem;
-class LifeTimeSystem;
-class LightSystem;
-class OctreeAddRemoverSystem;
-class MovementSystem;
-class MovementPostCollisionSystem;
-class CollisionSystem;
-class SpeedLimitSystem;
-class PrepareUpdateSystem;
-class GunSystem;
-class ProjectileSystem;
-class LevelGeneratorSystem;
-class GameInputSystem;
-class NetworkReceiverSystem;
-class NetworkSenderSystem;
-class AudioSystem;
-class RenderSystem;
+#include "Sail/entities/systems/SystemDeclarations.h"
 
 class NetworkSerializedPackageEvent;
 
@@ -70,31 +48,6 @@ private:
 	void initAnimations();
 
 private:
-	struct Systems {
-		AiSystem* aiSystem = nullptr;
-		AnimationSystem* animationSystem = nullptr;
-		CandleSystem* candleSystem = nullptr;
-		EntityAdderSystem* entityAdderSystem = nullptr;
-		EntityRemovalSystem* entityRemovalSystem = nullptr;
-		LifeTimeSystem* lifeTimeSystem = nullptr;
-		LightSystem* lightSystem = nullptr;
-		OctreeAddRemoverSystem* octreeAddRemoverSystem = nullptr;
-		UpdateBoundingBoxSystem* updateBoundingBoxSystem = nullptr;
-		PrepareUpdateSystem* prepareUpdateSystem = nullptr;
-		GunSystem* gunSystem = nullptr;
-		ProjectileSystem* projectileSystem = nullptr;
-		GameInputSystem* gameInputSystem = nullptr;
-		NetworkReceiverSystem* networkReceiverSystem = nullptr;
-		NetworkSenderSystem* networkSenderSystem = nullptr;
-		AudioSystem* audioSystem = nullptr;
-		RenderSystem* renderSystem = nullptr;
-		LevelGeneratorSystem* levelGeneratorSystem = nullptr;
-		MovementSystem* movementSystem = nullptr;
-		MovementPostCollisionSystem* movementPostCollisionSystem = nullptr;
-		CollisionSystem* collisionSystem = nullptr;
-		SpeedLimitSystem* speedLimitSystem = nullptr;
-	};
-
 	Application* m_app;
 	// Camera
 	PerspectiveCamera m_cam;
