+++ Sail/src/Sail/entities/systems/render/BeginEndFrameSystem.cpp
@@ -1,0 +1,15 @@
+#include "pch.h"
+#include "BeginEndFrameSystem.h"
+#include "..//..//..//Application.h"
+
+BeginEndFrameSystem::BeginEndFrameSystem() {
+}
+
+BeginEndFrameSystem::~BeginEndFrameSystem() {
+}
+
+void BeginEndFrameSystem::beginFrame(Camera& camera) {
+}
+
+void BeginEndFrameSystem::endFrameAndPresent() {
+}
