+++ Sail/src/Sail/entities/systems/render/BeginEndFrameSystem.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "..//BaseComponentSystem.h"
+#include "..//..//..//graphics/camera/Camera.h"
+
+class BeginEndFrameSystem final : public BaseComponentSystem {
+public:
+	BeginEndFrameSystem();
+	~BeginEndFrameSystem();
+
+	void beginFrame(Camera& camera);
+	void endFrameAndPresent();
+};
