--- Sail/src/Sail/entities/systems/render/EndFrameSystem.cpp
@@ -1,14 +1,0 @@
-#include "pch.h"
-#include "EndFrameSystem.h"
-
-EndFrameSystem::EndFrameSystem()
-{
-}
-
-EndFrameSystem::~EndFrameSystem()
-{
-}
-
-void EndFrameSystem::endAndPresentFrame()
-{
-}
