--- Sail/src/Sail/entities/systems/render/EndFrameSystem.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "..//BaseComponentSystem.h"
-
-class EndFrameSystem final : public BaseComponentSystem {
-public:
-	EndFrameSystem();
-	~EndFrameSystem();
-
-	void endAndPresentFrame();
-};
