+++ Sail/src/Sail/entities/systems/render/RenderSystem.cpp
@@ -70,8 +70,8 @@ void RenderSystem::draw(Camera& camera, const float alpha) {
 				Model* wireframeModel = boundingBox->getWireframeModel();
 				if (wireframeModel) {
 					// Bounding boxes are visualized with their most update since that's what's used for hit detection
-					//m_renderer->submit(wireframeModel, boundingBox->getTransform()->getMatrix(), Renderer::MESH_STATIC);
-					m_renderer->submit(wireframeModel, transform->getMatrix(), Renderer::MESH_STATIC);
+					m_renderer->submit(wireframeModel, boundingBox->getTransform()->getMatrix(), Renderer::MESH_STATIC);
+					//m_renderer->submit(wireframeModel, transform->getMatrix(), Renderer::MESH_STATIC);
 				}
 			}
 		}
